creation of MicroserviceRewards and Library
