+++ TourGuide/library/build.gradle
@@ -1,0 +1,27 @@
+plugins {
+    id 'java'
+}
+
+group 'com.tourguide.library'
+version '1.0'
+
+repositories {
+    mavenCentral()
+    flatDir {
+        dirs '../libs'
+    }
+}
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+
+    compile(name:'gpsUtil', ext:'jar')
+    compile(name:'TripPricer', ext:'jar')
+
+    implementation group: 'org.javamoney', name: 'moneta', version: '1.3'
+}
+
+test {
+    useJUnitPlatform()
+}
