+++ TourGuide/microservice-rewards/src/main/java/com/tourguide/microservices/rewards/MicroserviceRewards.java
@@ -1,0 +1,7 @@
+package com.tourguide.microservices.rewards;
+
+public class MicroserviceRewards {
+    public static void main(String[] args) {
+
+    }
+}
