+++ TourGuide/microservice-rewards/src/main/java/com/tourguide/microservices/rewards/module/TourGuideModule.java
@@ -1,0 +1,28 @@
+package com.tourguide.microservices.rewards.module;
+
+import gpsUtil.GpsUtil;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import rewardCentral.RewardCentral;
+
+
+@Configuration
+public class TourGuideModule {
+
+    @Bean
+    public GpsUtil getGpsUtil() {
+        return new GpsUtil();
+    }
+
+  /*  @Bean
+    public RewardsService getRewardsService() {
+        return new RewardsService(getGpsUtil(), getRewardCentral());
+    }*/
+
+    @Bean
+    public RewardCentral getRewardCentral() {
+        return new RewardCentral();
+    }
+
+}
+
