+++ TourGuide/microservice-rewards/src/main/java/com/tourguide/microservices/rewards/service/RewardsService.java
@@ -1,0 +1,77 @@
+package com.tourguide.microservices.rewards.service;
+
+import com.tourguide.library.user.User;
+import com.tourguide.library.user.UserReward;
+import gpsUtil.GpsUtil;
+import gpsUtil.location.Attraction;
+import gpsUtil.location.Location;
+import gpsUtil.location.VisitedLocation;
+import org.springframework.stereotype.Service;
+import rewardCentral.RewardCentral;
+
+import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
+
+
+@Service
+public class RewardsService {
+    private static final double STATUTE_MILES_PER_NAUTICAL_MILE = 1.15077945;
+
+    // proximity in miles
+    private final int defaultProximityBuffer = 10;
+    private int proximityBuffer = defaultProximityBuffer;
+    private final GpsUtil gpsUtil;
+    private final RewardCentral rewardsCentral;
+    private final List<Attraction> attractions = new CopyOnWriteArrayList<>();
+
+    public RewardsService(GpsUtil gpsUtil, RewardCentral rewardCentral) {
+        this.gpsUtil = gpsUtil;
+        this.rewardsCentral = rewardCentral;
+        attractions.addAll(gpsUtil.getAttractions());
+    }
+
+    public void setProximityBuffer(int proximityBuffer) {
+        this.proximityBuffer = proximityBuffer;
+    }
+
+    public void calculateRewards(User user) {
+        List<VisitedLocation> userLocations = user.getVisitedLocations();
+
+        for (VisitedLocation visitedLocation : userLocations) {
+            for (Attraction attraction : attractions) {
+                if (user.getUserRewards().stream().noneMatch(reward -> reward.attraction.attractionName.equals(attraction.attractionName))) {
+                    if (nearAttraction(visitedLocation, attraction)) {
+                        user.addUserReward(new UserReward(visitedLocation, attraction, getRewardPoints(attraction, user)));
+                    }
+                }
+            }
+        }
+    }
+
+    public boolean isWithinAttractionProximity(Attraction attraction, Location location) {
+        int attractionProximityRange = 200;
+        return !(getDistance(attraction, location) > attractionProximityRange);
+    }
+
+    private boolean nearAttraction(VisitedLocation visitedLocation, Attraction attraction) {
+        return !(getDistance(attraction, visitedLocation.location) > proximityBuffer);
+    }
+
+    public int getRewardPoints(Attraction attraction, User user) {
+        return rewardsCentral.getAttractionRewardPoints(attraction.attractionId, user.getUserId());
+    }
+
+    public double getDistance(Location loc1, Location loc2) {
+        double latitude1 = Math.toRadians(loc1.latitude);
+        double longitude1 = Math.toRadians(loc1.longitude);
+        double latitude2 = Math.toRadians(loc2.latitude);
+        double longitude2 = Math.toRadians(loc2.longitude);
+
+        double angle = Math.acos(Math.sin(latitude1) * Math.sin(latitude2)
+                + Math.cos(latitude1) * Math.cos(latitude2) * Math.cos(longitude1 - longitude2));
+
+        double nauticalMiles = 60 * Math.toDegrees(angle);
+        return STATUTE_MILES_PER_NAUTICAL_MILE * nauticalMiles;
+    }
+
+}
