+++ TourGuide/src/test/java/tourGuide/TestRewardsService.java
@@ -6,19 +6,16 @@ import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 
-import gpsUtil.location.Location;
-import org.junit.Ignore;
 import org.junit.Test;
 import gpsUtil.GpsUtil;
 import gpsUtil.location.Attraction;
 import gpsUtil.location.VisitedLocation;
 import rewardCentral.RewardCentral;
 import tourGuide.helper.InternalTestHelper;
-import tourGuide.model.UsersLocations;
 import tourGuide.service.RewardsService;
 import tourGuide.service.TourGuideService;
-import tourGuide.user.User;
-import tourGuide.user.UserReward;
+import com.tourguide.library.user.User;
+import com.tourguide.library.user.UserReward;
 
 
 public class TestRewardsService {
